pyproject.toml
[tool.poetry]
name = "estella"
version = "0.1.0"
description = ""
authors = ["du-cki <qwuakers@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.4.0"
aiohttp = "^3.9.5"
jishaku = "^2.5.2"
mcstatus = "^11.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
